implement clerk webhooks

+++ app/api/webhooks/clerk/route.ts
@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
-import { clerkClient } from "@clerk/nextjs";
+import { clerkClient } from "@clerk/clerk-sdk-node";
 import { WebhookEvent } from "@clerk/nextjs/server";
 import { headers } from "next/headers";
 import { NextResponse } from "next/server";
@@ -8,34 +8,38 @@ import { Webhook } from "svix";
 import { createUser, deleteUser, updateUser } from "@/lib/actions/user.actions";
 
 export async function POST(req: Request) {
-  // Get the webhook secret from environment variables
+  // You can find this in the Clerk Dashboard -> Webhooks -> choose the webhook
   const WEBHOOK_SECRET = process.env.WEBHOOK_SECRET;
 
   if (!WEBHOOK_SECRET) {
-    throw new Error("Please add WEBHOOK_SECRET from Clerk Dashboard to .env or .env.local");
+    throw new Error(
+      "Please add WEBHOOK_SECRET from Clerk Dashboard to .env or .env.local"
+    );
   }
 
-  // Get the necessary headers
+  // Get the headers
   const headerPayload = headers();
   const svix_id = headerPayload.get("svix-id");
   const svix_timestamp = headerPayload.get("svix-timestamp");
   const svix_signature = headerPayload.get("svix-signature");
 
-  // If the necessary headers are missing, return an error response
+  // If there are no headers, error out
   if (!svix_id || !svix_timestamp || !svix_signature) {
-    return new Response("Error occurred -- no svix headers", {
+    return new Response("Error occured -- no svix headers", {
       status: 400,
     });
   }
 
-  // Parse the request body
+  // Get the body
   const payload = await req.json();
   const body = JSON.stringify(payload);
 
-  // Verify the webhook using Svix
+  // Create a new Svix instance with your secret.
   const wh = new Webhook(WEBHOOK_SECRET);
+
   let evt: WebhookEvent;
 
+  // Verify the payload with the headers
   try {
     evt = wh.verify(body, {
       "svix-id": svix_id,
@@ -44,60 +48,68 @@ export async function POST(req: Request) {
     }) as WebhookEvent;
   } catch (err) {
     console.error("Error verifying webhook:", err);
-    return new Response("Error occurred", {
+    return new Response("Error occured", {
       status: 400,
     });
   }
 
-  // Handle the webhook event
+  // Get the ID and type
   const { id } = evt.data;
   const eventType = evt.type;
 
-  switch (eventType) {
-    case "user.created":
-      const { email_addresses, image_url, first_name, last_name, username } = evt.data;
+  // CREATE
+  if (eventType === "user.created") {
+    const { id, email_addresses, image_url, first_name, last_name, username } = evt.data;
 
-      const newUser = {
-        clerkId: id,
-        email: email_addresses[0].email_address,
-        username: username!,
-        firstName: first_name,
-        lastName: last_name,
-        photo: image_url,
-      };
+    const user = {
+      clerkId: id,
+      email: email_addresses[0].email_address,
+      username: username!,
+      firstName: first_name,
+      lastName: last_name,
+      photo: image_url,
+    };
 
-      const createdUser = await createUser(newUser);
+    const newUser = await createUser(user);
 
-      if (createdUser) {
-        await clerkClient.users.updateUserMetadata(id, {
-          publicMetadata: { userId: createdUser._id },
-        });
-      }
+    // Set public metadata
+    if (newUser) {
+      await clerkClient.users.updateUserMetadata(id, {
+        publicMetadata: {
+          userId: newUser._id,
+        },
+      });
+    }
 
-      return NextResponse.json({ message: "User created successfully", user: createdUser });
-
-    case "user.updated":
-      const updatedUserDetails = {
-        firstName: evt.data.first_name,
-        lastName: evt.data.last_name,
-        username: evt.data.username!,
-        photo: evt.data.image_url,
-      };
-
-      const updatedUser = await updateUser(id, updatedUserDetails);
-
-      return NextResponse.json({ message: "User updated successfully", user: updatedUser });
-
-    case "user.deleted":
-      const deletedUser = await deleteUser(id!);
-
-      return NextResponse.json({ message: "User deleted successfully", user: deletedUser });
-
-    default:
-      console.log(`Unhandled event type: ${eventType}`);
+    return NextResponse.json({ message: "OK", user: newUser });
   }
 
-  console.log(`Webhook with ID ${id} and type ${eventType}`);
+  // UPDATE
+  if (eventType === "user.updated") {
+    const { id, image_url, first_name, last_name, username } = evt.data;
+
+    const user = {
+      firstName: first_name,
+      lastName: last_name,
+      username: username!,
+      photo: image_url,
+    };
+
+    const updatedUser = await updateUser(id, user);
+
+    return NextResponse.json({ message: "OK", user: updatedUser });
+  }
+
+  // DELETE
+  if (eventType === "user.deleted") {
+    const { id } = evt.data;
+
+    const deletedUser = await deleteUser(id!);
+
+    return NextResponse.json({ message: "OK", user: deletedUser });
+  }
+
+  console.log(`Webhook with and ID of ${id} and type of ${eventType}`);
   console.log("Webhook body:", body);
 
   return new Response("", { status: 200 });
